moved logic in service, added Config for beans
