--- src/main/java/com/example/converter/Constants.java
@@ -1,5 +1,0 @@
-package com.example.converter;
-
-public class Constants {
-    public static final String URL_API = "https://www.cbr-xml-daily.ru/daily_json.js";
-}
