+++ src/main/java/com/example/converter/constants/Constants.java
@@ -1,0 +1,8 @@
+package com.example.converter.constants;
+
+public class Constants {
+    public static final String URL_API = "https://www.cbr-xml-daily.ru/daily_json.js";
+
+    // exceptions
+    public static final String EXC_PARAMS = "Params in the method are null";
+}
